testing: Remove dependency on the Go MySQL driver
