+++ consistent_test.go
@@ -14,12 +14,10 @@
 package parser
 
 import (
-	"bufio"
-	"fmt"
 	"io/ioutil"
-	"sort"
-	"strings"
 	"testing"
+
+	iextract "github.com/kyleconroy/sqlparse/internal/extract"
 
 	"github.com/google/go-cmp/cmp"
 )
@@ -42,93 +40,28 @@ func TestKeywordConsistent(t *testing.T) {
 	}
 
 	content := string(data)
-
-	reservedKeywords := extractKeywords(content, "ReservedKeyword")
-	unreservedKeywords := extractKeywords(content, "UnReservedKeyword")
-	notKeywordTokens := extractKeywords(content, "NotKeywordToken")
-	tidbKeywords := extractKeywords(content, "TiDBKeyword")
+	reservedKeywords := iextract.KeywordsFromTokens(content, iextract.KeywordReserved)
+	unreservedKeywords := iextract.KeywordsFromTokens(content, iextract.KeywordUnreserved)
+	notKeywordTokens := iextract.KeywordsFromTokens(content, iextract.KeywordNot)
+	tidbKeywords := iextract.KeywordsFromTokens(content, iextract.KeywordTiDB)
 
 	keywordCount := len(reservedKeywords) + len(unreservedKeywords) + len(notKeywordTokens) + len(tidbKeywords)
 	if diff := cmp.Diff(len(tokenMap)-len(aliases), keywordCount-len(windowFuncTokenMap)); diff != "" {
 		t.Errorf("length tokenMap does not match keyword count: %s", diff)
 	}
 
-	unreservedCollectionDef := extractKeywordsFromCollectionDef(content, "UnReservedKeyword:")
+	unreservedCollectionDef := iextract.KeywordsFromCollectionDef(content, iextract.KeywordUnreserved)
 	if diff := cmp.Diff(unreservedKeywords, unreservedCollectionDef); diff != "" {
 		t.Errorf("unreserved keywords: %s", diff)
 	}
 
-	notKeywordTokensCollectionDef := extractKeywordsFromCollectionDef(content, "NotKeywordToken:")
+	notKeywordTokensCollectionDef := iextract.KeywordsFromCollectionDef(content, iextract.KeywordNot)
 	if diff := cmp.Diff(notKeywordTokens, notKeywordTokensCollectionDef); diff != "" {
 		t.Errorf("not keyword tokens: %s", diff)
 	}
 
-	tidbKeywordsCollectionDef := extractKeywordsFromCollectionDef(content, "TiDBKeyword:")
+	tidbKeywordsCollectionDef := iextract.KeywordsFromCollectionDef(content, iextract.KeywordTiDB)
 	if diff := cmp.Diff(tidbKeywords, tidbKeywordsCollectionDef); diff != "" {
 		t.Errorf("TiDB keywords: %s", diff)
 	}
-}
-
-func extractMiddle(str, startMarker, endMarker string) string {
-	startIdx := strings.Index(str, startMarker)
-	if startIdx == -1 {
-		return ""
-	}
-	str = str[startIdx+len(startMarker):]
-	endIdx := strings.Index(str, endMarker)
-	if endIdx == -1 {
-		return ""
-	}
-	return str[:endIdx]
-}
-
-func extractLines(str, startMarker, endMarker string) []string {
-	var started, stopped bool
-	var lines []string
-	scanner := bufio.NewScanner(strings.NewReader(str))
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-		if line == startMarker {
-			started = true
-			continue
-		}
-		if started && line == endMarker {
-			stopped = true
-			continue
-		}
-		if stopped {
-			continue
-		}
-		if started {
-			lines = append(lines, line)
-		}
-	}
-	if err := scanner.Err(); err != nil {
-		panic(err)
-	}
-	return lines
-}
-
-func extractQuotedWords(strs []string) []string {
-	var words []string
-	for _, str := range strs {
-		word := extractMiddle(str, "\"", "\"")
-		if word == "" {
-			continue
-		}
-		words = append(words, word)
-	}
-	sort.Strings(words)
-	return words
-}
-
-func extractKeywords(content, keyword string) []string {
-	start := fmt.Sprintf("/* The following tokens belong to %s. Notice: make sure these tokens are contained in %s. */", keyword, keyword)
-	lines := extractLines(content, start, "")
-	return extractQuotedWords(lines)
-}
-
-func extractKeywordsFromCollectionDef(content, startMarker string) []string {
-	lines := extractLines(content, startMarker, "")
-	return extractQuotedWords(lines)
 }
