+++ go.mod
@@ -1,7 +1,6 @@
 module github.com/kyleconroy/sqlparse
 
 require (
-	github.com/go-sql-driver/mysql v1.5.0
 	github.com/google/go-cmp v0.5.4
 	github.com/pingcap/check v0.0.0-20190102082844-67f458068fc8
 	github.com/pingcap/errors v0.11.5-0.20201029093017-5a7df2af2ac7
